Add a ton of documentation
