+++ SmartDeviceLink/NSMutableDictionary+SafeRemove.h
@@ -10,12 +10,26 @@
 
 @interface NSMutableDictionary <KeyType, ObjectType> (SafeRemove)
 
+/**
+ *  Safely attempt to remove a key/object pair by no-oping if they don't exist in the Mutable Dictionary instead of throwing an exception
+ *
+ *  @param aKey The key to attempt removal of
+ *
+ *  @return Whether or not the key/object pair existed to remove.
+ */
 - (BOOL)safeRemoveObjectForKey:(KeyType)aKey;
 
 @end
 
 @interface NSMapTable <KeyType, ObjectType> (SafeRemove)
 
+/**
+ *  Safely attempt to remove a key/object pair by no-oping if they don't exist in the MapTable instead of throwing an exception
+ *
+ *  @param aKey The key to attempt removal of
+ *
+ *  @return Whether or not the key/object pair existed to remove.
+ */
 - (BOOL)safeRemoveObjectForKey:(KeyType)aKey;
 
 @end
