+++ SmartDeviceLink/NSNumber+NumberType.h
@@ -8,15 +8,27 @@
 
 #import <Foundation/Foundation.h>
 
+/**
+ *  A decleration that this NSNumber contains an NSInteger.
+ */
 @protocol SDLInt
 @end
 
+/**
+ *  A declaration that this NSNumber contains an NSUInteger.
+ */
 @protocol SDLUInt
 @end
 
+/**
+ *  A declaration that this NSNumber contains a BOOL.
+ */
 @protocol SDLBool
 @end
 
+/**
+ *  A declaration that this NSNumber contains a float.
+ */
 @protocol SDLFloat
 @end
