+++ SmartDeviceLink/SDLDeleteFileOperation.h
@@ -17,6 +17,15 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SDLDeleteFileOperation : NSOperation
 
+/**
+ *  Create an instance of a delete files operation which will tell the remote system to remove a file form its storage.
+ *
+ *  @param fileName The name of the file to be deleted on the remote system.
+ *  @param connectionManager The connection manager which will handle transporting the request to the remote system.
+ *  @param completionHandler A completion handler to be called when the delete finishes.
+ *
+ *  @return An instance of SDLDeleteFilesOperation
+ */
 - (instancetype)initWithFileName:(NSString *)fileName connectionManager:(id<SDLConnectionManagerType>)connectionManager completionHandler:(nullable SDLFileManagerDeleteCompletion)completionHandler;
 
 @end
