+++ SmartDeviceLink/SDLPermissionConstants.h
@@ -16,15 +16,39 @@ NS_ASSUME_NONNULL_BEGIN
 typedef NSString SDLPermissionRPCName;
 typedef NSUUID SDLPermissionObserverIdentifier;
 
+/**
+ *  A permission group type which will be used to tell the system what type of changes you want to be notified about for the group.
+ */
 typedef NS_ENUM(NSUInteger, SDLPermissionGroupType) {
+    /**
+     *  Be notified when all of the RPC in the group are allowed, or, when they all stop being allowed in some sense, that is, when they were all allowed, and now they are not.
+     */
     SDLPermissionGroupTypeAllAllowed,
+    /**
+     *  Be notified when any change in availability occurs among the group.
+     */
     SDLPermissionGroupTypeAny
 };
 
+/**
+ *  The status of the group of RPCs permissions.
+ */
 typedef NS_ENUM(NSUInteger, SDLPermissionGroupStatus) {
+    /**
+     *  Every RPC in the group is currently allowed.
+     */
     SDLPermissionGroupStatusAllowed,
+    /**
+     *  Every RPC in the group is currently disallowed.
+     */
     SDLPermissionGroupStatusDisallowed,
+    /**
+     *  Some RPCs in the group are allowed and some disallowed.
+     */
     SDLPermissionGroupStatusMixed,
+    /**
+     *  The current status of the group is unknown.
+     */
     SDLPermissionGroupStatusUnknown
 };
