+++ SmartDeviceLink/SDLPermissionFilter.h
@@ -15,15 +15,55 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SDLPermissionFilter : NSObject <NSCopying>
 
+/**
+ *  An identifier for the permission filter to allow it to be removed at a later time.
+ */
 @property (copy, nonatomic, readonly) SDLPermissionObserverIdentifier *identifier;
+
+/**
+ *  All of the RPC names in this filter group.
+ */
 @property (copy, nonatomic, readonly) NSArray<SDLPermissionRPCName *> *rpcNames;
+
+/**
+ *  The type of this filter group.
+ */
 @property (assign, nonatomic, readonly) SDLPermissionGroupType groupType;
+
+/**
+ *  The block that will be called on status changes to this filter group.
+ */
 @property (copy, nonatomic, readonly) SDLPermissionObserver observer;
 
+/**
+ *  Create a new permission filter group.
+ *
+ *  @param rpcNames  The names of the RPCs to watch permissions of.
+ *  @param groupType The type of notifications to be sent for this filter group.
+ *  @param observer  The block observer to be called when changes occur.
+ *
+ *  @return An instance of `SDLPermissionFilter`.
+ */
 - (instancetype)initWithRPCNames:(NSArray<SDLPermissionRPCName *> *)rpcNames groupType:(SDLPermissionGroupType)groupType observer:(SDLPermissionObserver)observer NS_DESIGNATED_INITIALIZER;
 
+/**
+ *  Create a new permission filter group.
+ *
+ *  @param rpcNames  The names of the RPCs to watch permissions of.
+ *  @param groupType The type of notifications to be sent for this filter group.
+ *  @param observer  The block observer to be called when changes occur.
+ *
+ *  @return An instance of `SDLPermissionFilter`.
+ */
 + (instancetype)filterWithRPCNames:(NSArray<SDLPermissionRPCName *> *)rpcNames groupType:(SDLPermissionGroupType)groupType observer:(SDLPermissionObserver)observer;
 
+/**
+ *  Whether the current filter is equivalent with another filter or not.
+ *
+ *  @param otherFilter The other filter to test equivalence with.
+ *
+ *  @return YES if this filter is equivalent with `otherFilter`, NO otherwise.
+ */
 - (BOOL)isEqualToFilter:(SDLPermissionFilter *)otherFilter;
 
 @end
