+++ SmartDeviceLink/SDLRequestHandler.h
@@ -14,8 +14,18 @@ NS_ASSUME_NONNULL_BEGIN
 
 @protocol SDLRequestHandler <NSObject>
 
+/**
+ *  The handler that is added to any RPC implementing this protocol.
+ */
 @property (copy, nonatomic, readonly) SDLRPCNotificationHandler handler;
 
+/**
+ *  A special init function on any RPC implementing this protocol.
+ *
+ *  @param handler The handler to be called at specified times.
+ *
+ *  @return An instance of the class implementing this protocol.
+ */
 - (instancetype)initWithHandler:(SDLRPCNotificationHandler)handler;
 
 @end
