+++ SmartDeviceLink/SDLUploadFileOperation.h
@@ -19,6 +19,14 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SDLUploadFileOperation : NSOperation
 
+/**
+ *  Create an instance of an upload files operation which will send a file to a remote system when added to an operation queue.
+ *
+ *  @param file A file wrapper around the file which will be sent and a completion handler for when the file finishes sending.
+ *  @param connectionManager The connection manager which will handle transporting the file bytes to the remote system
+ *
+ *  @return An instance of SDLUploadFilesOperation
+ */
 - (instancetype)initWithFile:(SDLFileWrapper *)file connectionManager:(id<SDLConnectionManagerType>)connectionManager;
 
 @end
